requirements.txt
flask==2.3.3
flask_sqlalchemy==3.1.1
flask_login==0.6.2
werkzeug==2.3.7
requests==2.31.0
pycoingecko==3.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
plotly==5.17.0
pandas==2.1.1


